(feat): Implements ProjectConfig interface and creates DatafileProjectConfig. (#177)
