+++ lib/optimizely/audience.rb
@@ -24,7 +24,7 @@ module Optimizely
   module Audience
     module_function
 
-    def user_in_experiment?(config, experiment, attributes)
+    def user_in_experiment?(config, experiment, attributes, logger)
       # Determine for given experiment if user satisfies the audiences for the experiment.
       #
       # config - Representation of the Optimizely project config.
@@ -36,7 +36,7 @@ module Optimizely
 
       audience_conditions = experiment['audienceConditions'] || experiment['audienceIds']
 
-      config.logger.log(
+      logger.log(
         Logger::DEBUG,
         format(
           Helpers::Constants::AUDIENCE_EVALUATION_LOGS['EVALUATING_AUDIENCES_COMBINED'],
@@ -47,7 +47,7 @@ module Optimizely
 
       # Return true if there are no audiences
       if audience_conditions.empty?
-        config.logger.log(
+        logger.log(
           Logger::INFO,
           format(
             Helpers::Constants::AUDIENCE_EVALUATION_LOGS['AUDIENCE_EVALUATION_RESULT_COMBINED'],
@@ -60,7 +60,7 @@ module Optimizely
 
       attributes ||= {}
 
-      custom_attr_condition_evaluator = CustomAttributeConditionEvaluator.new(attributes, config.logger)
+      custom_attr_condition_evaluator = CustomAttributeConditionEvaluator.new(attributes, logger)
 
       evaluate_custom_attr = lambda do |condition|
         return custom_attr_condition_evaluator.evaluate(condition)
@@ -71,7 +71,7 @@ module Optimizely
         return nil unless audience
 
         audience_conditions = audience['conditions']
-        config.logger.log(
+        logger.log(
           Logger::DEBUG,
           format(
             Helpers::Constants::AUDIENCE_EVALUATION_LOGS['EVALUATING_AUDIENCE'],
@@ -83,7 +83,7 @@ module Optimizely
         audience_conditions = JSON.parse(audience_conditions) if audience_conditions.is_a?(String)
         result = ConditionTreeEvaluator.evaluate(audience_conditions, evaluate_custom_attr)
         result_str = result.nil? ? 'UNKNOWN' : result.to_s.upcase
-        config.logger.log(
+        logger.log(
           Logger::INFO,
           format(Helpers::Constants::AUDIENCE_EVALUATION_LOGS['AUDIENCE_EVALUATION_RESULT'], audience_id, result_str)
         )
@@ -94,7 +94,7 @@ module Optimizely
 
       eval_result ||= false
 
-      config.logger.log(
+      logger.log(
         Logger::INFO,
         format(
           Helpers::Constants::AUDIENCE_EVALUATION_LOGS['AUDIENCE_EVALUATION_RESULT_COMBINED'],
