+++ lib/optimizely/decision_service.rb
@@ -94,7 +94,7 @@ module Optimizely
       end
 
       # Check audience conditions
-      unless Audience.user_in_experiment?(project_config, experiment, attributes)
+      unless Audience.user_in_experiment?(project_config, experiment, attributes, @logger)
         @logger.log(
           Logger::INFO,
           "User '#{user_id}' does not meet the conditions to be in experiment '#{experiment_key}'."
@@ -236,7 +236,7 @@ module Optimizely
         audience_name = audience['name']
 
         # Check that user meets audience conditions for targeting rule
-        unless Audience.user_in_experiment?(project_config, rollout_rule, attributes)
+        unless Audience.user_in_experiment?(project_config, rollout_rule, attributes, @logger)
           @logger.log(
             Logger::DEBUG,
             "User '#{user_id}' does not meet the conditions to be in rollout rule for audience '#{audience_name}'."
@@ -255,7 +255,7 @@ module Optimizely
       # get last rule which is the everyone else rule
       everyone_else_experiment = rollout_rules[number_of_rules]
       # Check that user meets audience conditions for last rule
-      unless Audience.user_in_experiment?(project_config, everyone_else_experiment, attributes)
+      unless Audience.user_in_experiment?(project_config, everyone_else_experiment, attributes, @logger)
         audience_id = everyone_else_experiment['audienceIds'][0]
         audience = project_config.get_audience_from_id(audience_id)
         audience_name = audience['name']
